Move reminder views to rmndin.reminders.views
